+++ src/test/java/com/together/buytogether/post/feature/RegisterPostTest.java
@@ -8,10 +8,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import org.springframework.util.Assert;
 
 import java.time.LocalDateTime;
-import java.util.*;
+import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -52,108 +51,8 @@ public class RegisterPostTest {
                 expiredAt
         );
         registerPost.requet(request);
-        assertThat(postRepository.findAll()).isEqualTo(1);
+        assertThat(postRepository.findAll()).hasSize(1);
 
     }
 
-    public static class Post {
-        private final Member member;
-        private final String title;
-        private final String content;
-        private final LocalDateTime expiredAt;
-        private Long id;
-
-        public Post(
-                Member member,
-                String title,
-                String content,
-                LocalDateTime expiredAt) {
-            validateConstructor(
-                    member,
-                    title,
-                    content,
-                    expiredAt);
-
-            this.member = member;
-            this.title = title;
-            this.content = content;
-            this.expiredAt = expiredAt;
-
-
-        }
-
-        private static void validateConstructor(
-                Member member,
-                String title,
-                String content,
-                LocalDateTime expiredAt) {
-            Assert.notNull(member, "회원 정보는 필수입니다");
-            Assert.hasText(title, "글 제목은 필수입니다");
-            Assert.hasText(content, "글 내용은 필수입니다");
-            Assert.notNull(expiredAt, "글 만료일은 필수입니다");
-        }
-
-        public Long getId() {
-            return id;
-        }
-
-        public void assingId(Long id) {
-            this.id = id;
-        }
-    }
-
-    private class RegisterPost {
-        private final MemberRepository memberRepository;
-        private final PostRepository postRepository;
-
-        public RegisterPost(PostRepository postRepository, MemberRepository memberRepository) {
-            this.memberRepository = memberRepository;
-            this.postRepository = postRepository;
-        }
-
-        public void requet(Request request) {
-            Member member = memberRepository.findById(request.memberId).orElseThrow(
-                    () -> new IllegalArgumentException("존재하지 않는 회원입니다"));
-            Post post = request.toDomain(member);
-            postRepository.save(post);
-        }
-
-        public record Request(
-                Long memberId,
-                String title,
-                String content,
-                LocalDateTime expiredAt) {
-            public Request {
-                Assert.notNull(memberId, "회원 번호는 필수입니다");
-                Assert.hasText(title, "글 제목은 필수입니다");
-                Assert.hasText(content, "글 내용은 필수입니다");
-                Assert.notNull(expiredAt, "글 만료일은 필수입니다");
-            }
-
-
-            public Post toDomain(Member member) {
-
-                return new Post(
-                        member,
-                        title,
-                        content,
-                        expiredAt
-                );
-            }
-        }
-    }
-
-    private class PostRepository {
-        private final Map<Long, Post> posts = new HashMap<>();
-        private Long id = 1L;
-
-        public void save(Post post) {
-            post.assingId(id++);
-            posts.put(post.getId(), post);
-        }
-
-        public List<Post> findAll() {
-            return new ArrayList<>(posts.values());
-        }
-    }
 }
